+++ classes/utils/Asynchronzier.php
@@ -252,6 +252,7 @@ class Asynchronzier
                                 unset($localLines[$index]);
                             }
                         }
+                        // dd($estimateObj->Line);
                         foreach ($estimateObj->Line as $lineObj) {
                             $parsedLine = ERPDataParser::parseEstimateLine($lineObj, $estimateObj->Id);
                             if ($parsedLine) {
@@ -533,6 +534,7 @@ class Asynchronzier
     public function Update($entity)
     {
         $object = new $entity['name']();
+        $object->sparse = true;
         foreach ($entity['attributes'] as $key => $value) {
             if (!is_array($entity['attributes'][$key])) {
                 $object->$key = $value;
@@ -714,11 +716,11 @@ class Asynchronzier
                     [
                         'name' => 'IPPPhysicalAddress',
                         'attributes' => [
-                            'Id' => $localData['bill_address_id'],
-                            'Line1' => $localData['bill_address'],
-                            'City' => $localData['bill_city'],
-                            'CountrySubDivisionCode' => $localData['bill_state'],
-                            'PostalCode' => $localData['bill_zip_code']
+                            'Id' => 12,
+                            // 'Line1' => $localData['bill_address'],
+                            // 'City' => $localData['bill_city'],
+                            // 'CountrySubDivisionCode' => $localData['bill_state'],
+                            // 'PostalCode' => $localData['bill_zip_code']
                         ],
                     ],
                 ],