+++ js/app/estimate/AddEstimateCtrl.js
@@ -150,6 +150,7 @@ function AddEstimateCtrl(
                         $scope.estimate.bill_city = cus.bill_city;
                         $scope.estimate.bill_state = cus.bill_state;
                         $scope.estimate.bill_zip_code = cus.bill_zip_code;
+                        $scope.estimate.bill_country = cus.bill_country;
                         $scope.estimate.primary_phone_number = cus.primary_phone_number;
                         $scope.estimate.alternate_phone_number = cus.alternate_phone_number;
                         $scope.estimate.email = cus.email;
@@ -170,6 +171,7 @@ function AddEstimateCtrl(
                         $scope.estimate.job_city = cus.ship_city;
                         $scope.estimate.job_state = cus.ship_state;
                         $scope.estimate.job_zip_code = cus.ship_zip_code;
+                        $scope.estimate.job_country = cus.ship_country;
                     }
                     $scope.estimate.job_customer_display_name = cus.display_name;
                     return;